refactor(bpane-host): extract session, message_dispatch, and test_session; finalize main.rs
Complete the main.rs refactoring by extracting the remaining components:

- session.rs (266 lines): FFmpeg session orchestration — spawns all
  subsystem tasks, NAL bridge, tile thread, and message dispatch
- message_dispatch/ (243+118 lines): gateway message select loop with
  control, input, clipboard, audio, video, and file transfer handling
- test_session/ (120+85 lines): fallback test session with test backends

main.rs is now 224 lines — a clean entry point with CLI parsing,
display/audio/camera detection, IPC binding, and session routing.
Down 94% from the original 3842 lines. All 287 tests pass.
